pyproject.toml
[project]
name = "g-maker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=24.1.0",
    "openai>=1.86.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "yt-dlp>=2025.6.9",
]
